cpt_generations/requirements.txt
huggingface_hub
pandas
python-dotenv
scikit_learn
torch
tqdm
transformers
numpy
